requirements.txt
a2wsgi
agate
aiofiles
aiohappyeyeballs
aiohttp
aiologic
aiosignal
aiosmtplib
aiosqlite
alembic
annotated-types
anyio
apache-airflow
apache-airflow-core
apache-airflow-providers-cncf-kubernetes
apache-airflow-providers-common-compat
apache-airflow-providers-common-io
apache-airflow-providers-common-sql
apache-airflow-providers-smtp
apache-airflow-providers-standard
apache-airflow-task-sdk
argcomplete
asgiref
attrs
awscli
babel
backports.strenum
backports.tarfile
blinker
botocore
cachetools
cadwyn
certifi
cffi
charset-normalizer
click
colorama
colorlog
cron_descriptor
croniter
cryptography
daff
databricks-sdk
databricks-sql-connector
dbt-adapters
dbt-common
dbt-core
dbt-extractor
dbt-protos
dbt-semantic-interfaces
deepdiff
Deprecated
dill
dnspython
docutils
durationpy
email-validator
et_xmlfile
eval_type_backport
exceptiongroup
fastapi
fastapi-cli
Flask
frozenlist
fsspec
google-auth
googleapis-common-protos
graphviz
greenback
greenlet
grpcio
gunicorn
h11
httpcore
httptools
httpx
idna
importlib_metadata
isodate
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
Jinja2
jmespath
jsonschema
jsonschema-specifications
keyring
kubernetes
kubernetes_asyncio
lazy-object-proxy
leather
libcst
linkify-it-py
lockfile
lz4
Mako
markdown-it-py
MarkupSafe
mashumaro
mdurl
methodtools
more-itertools
msgpack
msgspec
multidict
natsort
networkx
numpy
oauthlib
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
orderly-set
outcome
packaging
pandas
parsedatetime
pathlib_abc
pathspec
pendulum
pluggy
propcache
protobuf
psutil
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pygtrie
PyJWT
python-daemon
python-dateutil
python-dotenv
python-multipart
python-slugify
pytimeparse
pytz
PyYAML
PyYAML-ft
referencing
requests
requests-oauthlib
retryhttp
rich
rich-argparse
rich-toolkit
rpds-py
rsa
s3transfer
setproctitle
shellingham
six
sniffio
snowplow-tracker
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparams
sqlparse
starlette
structlog
svcs
tabulate
tenacity
termcolor
text-unidecode
thrift
tomli
typer
types-requests
types-urllib3
typing-inspection
typing_extensions
tzdata
uc-micro-py
universal_pathlib
urllib3
uuid6
uv
uvicorn
uvloop
watchfiles
websocket-client
websockets
Werkzeug
wirerope
wrapt
yarl
zipp